# [Neural Speed] Enable StableLM2-1.6B & StableLM2-Zephyr-1.6B & StableLM-3B