--- a/neural_speed/application/main_pybind.cpp
+++ b/neural_speed/application/main_pybind.cpp
@@ -896,6 +896,10 @@ PYBIND11_MODULE(phi_cpp, m)
 
 #elif MODEL_NAME_ID == 17
 
+PYBIND11_MODULE(stablelm_cpp, m)
+
+#elif MODEL_NAME_ID == 18
+
 PYBIND11_MODULE(whisper_cpp, m)
 
 #endif