--- a/neural_speed/convert/convert_stablelm.py
+++ b/neural_speed/convert/convert_stablelm.py
@@ -19,7 +19,7 @@
 #
 # This script is similar to "convert-pt-to-ne.py"
 #
-
+import os
 import struct
 import numpy as np
 from pathlib import Path
@@ -74,14 +74,14 @@ def bpe(mergeable_ranks, token, max_rank):
 def stablelm_convert_gguf(model, tokenizer, dir_model, fname_out, ftype, hparams):
     print("stablelm.gguf converting: ")
     list_vars = model.state_dict()
-    n_rot = int(hparams["partial_rotary_factor"]*hparams["hidden_size"]/hparams["num_attention_heads"])
+    n_rot = int(hparams["partial_rotary_factor"] * hparams["hidden_size"] / hparams["num_attention_heads"])
     for name in list_vars.keys():
         print(name, list_vars[name].shape, list_vars[name].dtype)
 
     print(hparams)
 
     gguf_file = fname_out + '.gguf'
-    gguf_writer = gguf.GGUFWriter(gguf_file, "phi")
+    gguf_writer = gguf.GGUFWriter(gguf_file, "stablelm")
 
     gguf_writer.add_uint32('magic', 0x67676d66)
     gguf_writer.add_uint32('version', 1)
@@ -96,10 +96,10 @@ def stablelm_convert_gguf(model, tokenizer, dir_model, fname_out, ftype, hparams
     gguf_writer.add_context_length(hparams["max_position_embeddings"])
     gguf_writer.add_feed_forward_length(hparams["intermediate_size"])
 
-    gguf_writer.add_bos_token_id(tokenizer.bos_token_id)
-    gguf_writer.add_eos_token_id(tokenizer.eos_token_id)
-    gguf_writer.add_pad_token_id(tokenizer.pad_token_id if tokenizer.pad_token_id is not None else 0)
-    gguf_writer.add_sep_token_id(tokenizer.sep_token_id if tokenizer.sep_token_id is not None else 0)
+    gguf_writer.add_bos_token_id(hparams["bos_token_id"])
+    gguf_writer.add_eos_token_id(hparams["eos_token_id"])
+    gguf_writer.add_pad_token_id(hparams["pad_token_id"] if hparams["pad_token_id"] else 0)
+    gguf_writer.add_sep_token_id(hparams["sep_token_id"] if hparams["sep_token_id"] else 0)
 
     def write_vocab_gguf(dir_model, hparams, gguf_writer):
         tokens: list[bytearray] = []
@@ -175,15 +175,16 @@ def stablelm_convert_gguf(model, tokenizer, dir_model, fname_out, ftype, hparams
 
         special_vocab = gguf.SpecialVocab(dir_model, load_merges=False)
         special_vocab.merges = merges
-        special_vocab._set_special_token("bos", tokenizer.special_tokens["<|endoftext|>"])
-        special_vocab._set_special_token("eos", tokenizer.special_tokens["<|endoftext|>"])
+        if not special_vocab.special_token_ids:
+            special_vocab._set_special_token("bos", tokenizer.special_tokens["<|endoftext|>"])
+            special_vocab._set_special_token("eos", tokenizer.special_tokens["<|endoftext|>"])
         special_vocab._set_special_token("unk", tokenizer.special_tokens["<|endoftext|>"])
         special_vocab.add_to_gguf(gguf_writer)
 
-
     # StableLM2 1.6B is different from StableLM3B, by having a special tiktoken tokenizer,
-    # so they don't have a tokenizer.json
-    if (dir_model / "tokenizer.json").is_file():
+    # so we detect how to write vocab based on the presence of a tokenizer.json
+    tokenizer_path = os.path.join(dir_model, "tokenizer.json")
+    if os.path.exists(tokenizer_path):
         write_vocab_gguf(dir_model, hparams, gguf_writer)
     else:
         write_vocab_gguf_stablelm2(dir_model, hparams, gguf_writer)
@@ -217,9 +218,6 @@ def stablelm_convert_gguf(model, tokenizer, dir_model, fname_out, ftype, hparams
                 data = data.astype(np.float32)
                 ftype_cur = 0
 
-        # print(f"[{i+1:{padi}d}/{len(model)}]
-        # Writing tensor {name:38s} | size {size:16} | type {lazy_tensor.data_type.name:4}")
-
         gguf_writer.add_tensor(name, data)
 
     print("gguf: write header")
@@ -234,12 +232,13 @@ def stablelm_convert_gguf(model, tokenizer, dir_model, fname_out, ftype, hparams
     print("Done. Output file: " + fname_out)
     print("")
 
-def phi_convert(model, tokenizer, dir_model, fname_out, ftype, hparams):
-    n_rot = int(hparams["partial_rotary_factor"]*hparams["hidden_size"]/hparams["num_attention_heads"])
+def stablelm_convert(model, tokenizer, dir_model, fname_out, ftype, hparams):
+    n_rot = int(hparams["partial_rotary_factor"] * hparams["hidden_size"] / hparams["num_attention_heads"])
     model.eval()
     for p in model.parameters():
         p.requires_grad = False
     hparams = model.config.to_dict()
+    vocab_size = hparams["vocab_size"]
     print("Model loaded: ", dir_model)
 
     fout = open(fname_out, "wb")
@@ -269,25 +268,59 @@ def phi_convert(model, tokenizer, dir_model, fname_out, ftype, hparams):
     fout.write(struct.pack("i", 0))
     fout.write(struct.pack("i", 0))
     fout.write(struct.pack("i", 0))
-    fout.write(struct.pack("f", hparams.get("rms_norm_eps", 1e-6)))  # rms norm eps
-    fout.write(struct.pack("f", 10000.0))  # freq_base
-    fout.write(struct.pack("f", 1.0))  # rope_factor
-    fout.write(struct.pack("i", tokenizer.bos_token_id if tokenizer.bos_token_id is not None else -1))
-    fout.write(struct.pack("i", tokenizer.eos_token_id if tokenizer.eos_token_id is not None else -1))
-    fout.write(struct.pack("i", tokenizer.pad_token_id if tokenizer.pad_token_id is not None else -1))
-    fout.write(struct.pack("i", tokenizer.sep_token_id if tokenizer.sep_token_id is not None else -1))
+    fout.write(struct.pack("f", hparams["layer_norm_eps"]))  # rms norm eps
+    fout.write(struct.pack("f", hparams["rope_theta"]))  # freq_base
+    fout.write(struct.pack("f", hparams["partial_rotary_factor"]))  # rope_factor
+    fout.write(struct.pack("i", hparams["bos_token_id"]))
+    fout.write(struct.pack("i", hparams["eos_token_id"]))
+    fout.write(struct.pack("i", hparams["pad_token_id"] if hparams["pad_token_id"] else 0))
+    fout.write(struct.pack("i", hparams["sep_token_id"] if hparams["sep_token_id"] else 0))
 
-    for i in range(hparams["vocab_size"]):
-        if i < tokenizer.vocab_size:
-            text = tokenizer.decode([i]).encode('utf-8')
-            fout.write(struct.pack("i", len(text)))
-            fout.write(text)
-            fout.write(struct.pack("f", 0.0 - i))
-        else:
-            text = tokenizer.decode([tokenizer.vocab_size - 1]).encode('utf-8')
-            fout.write(struct.pack("i", len(text)))
-            fout.write(text)
-            fout.write(struct.pack("f", -10000))
+    tokenizer_path = os.path.join(dir_model, "tokenizer.json")
+    print(tokenizer_path)
+    print(os.path.exists(tokenizer_path))
+
+    # For StableLM-3B
+    if os.path.exists(tokenizer_path):
+        for i in range(vocab_size):
+            if i < vocab_size:
+                text = tokenizer.decode([i]).encode('utf-8')
+                fout.write(struct.pack("i", len(text)))
+                fout.write(text)
+                fout.write(struct.pack("f", 0.0 - i))
+            else:
+                text = tokenizer.decode([vocab_size - 1]).encode('utf-8')
+                fout.write(struct.pack("i", len(text)))
+                fout.write(text)
+                fout.write(struct.pack("f", -10000))
+
+    # For StableLM2-1.6B & StableLM2-Zephyr-1.6B
+    else:
+        merges = []
+        vocab = {}
+        mergeable_ranks = tokenizer.mergeable_ranks
+        for token, rank in mergeable_ranks.items():
+            vocab[token_bytes_to_string(token)] = rank
+            if len(token) == 1:
+                continue
+            merged = bpe(mergeable_ranks, token, max_rank=rank)
+            assert len(merged) == 2
+            merges.append(' '.join(map(token_bytes_to_string, merged)))
+
+        added_vocab = tokenizer.special_tokens
+        reverse_vocab = {id_ : encoded_tok for encoded_tok, id_ in (vocab | added_vocab).items()}
+
+        for i in range(vocab_size):
+            if i in reverse_vocab:
+                text = reverse_vocab[i].encode('utf-8')
+                fout.write(struct.pack("i", len(text)))
+                fout.write(text)
+                fout.write(struct.pack("f", 0.0 - i))
+            else:
+                text = f"[PAD{i}]".encode("utf-8")
+                fout.write(struct.pack("i", len(text)))
+                fout.write(text)
+                fout.write(struct.pack("f", -10000))
 
     list_vars = model.state_dict()
 
@@ -331,15 +364,30 @@ def phi_convert(model, tokenizer, dir_model, fname_out, ftype, hparams):
     print("")
 
 def main(args_in: Optional[List[str]] = None) -> None:
-    parser = argparse.ArgumentParser(description="Convert a model to a NE compatible file")
-    parser.add_argument("--outtype", choices=["f32", "f16"], help="output format (default: based on input)")
-    parser.add_argument("--outfile", type=Path, help="path to write to; default: based on input")
-    parser.add_argument("model", type=Path, help="directory containing model file")
-    parser.add_argument("--format",
-                        type=str,
-                        default="NE",
-                        choices=["NE", "GGUF"],
-                        help="convert to the GGUF or NE format")
+    parser = argparse.ArgumentParser(description="Convert a model to an NE or GGUF compatible file")
+    parser.add_argument(
+        "--outtype",
+        choices=["f32", "f16"], 
+        help="output format (default: based on input)"
+    )
+    parser.add_argument(
+        "--outfile",
+        type=Path,
+        help="path to write to; default: based on input"
+    )
+    parser.add_argument(
+        "--format",
+        type=str,
+        default="NE",
+        choices=["NE", "GGUF"],
+        help="convert to the GGUF or NE format"
+    )
+    parser.add_argument(
+        "model",
+        type=Path,
+        help="directory containing model file"
+    )
+
     args = parser.parse_args(args_in)
 
     dir_model = args.model.as_posix()
@@ -357,9 +405,9 @@ def main(args_in: Optional[List[str]] = None) -> None:
     model = AutoModelForCausalLM.from_pretrained(dir_model, trust_remote_code=True)
     hparams = model.config.to_dict()
     if args.format == "GGUF":
-        phi_convert_gguf(model, tokenizer, dir_model, fname_out, ftype, hparams)
+        stablelm_convert_gguf(model, tokenizer, dir_model, fname_out, ftype, hparams)
     else:
-        phi_convert(model, tokenizer, dir_model, fname_out, ftype, hparams)
+        stablelm_convert(model, tokenizer, dir_model, fname_out, ftype, hparams)