… docs/supported_models.md
… neural_speed/application/CMakeLists.txt
… neural_speed/core/ne_layers.c
… neural_speed/models/CMakeLists.txt
… neural_speed/models/model_utils/model_types.h
… neural_speed/models/stablelm/stablelm.cpp
… neural_speed/models/stablelm/stablelm.h
… neural_speed/models/stablelm/stablelm_utils.cpp
… scripts/convert.py
… scripts/run.py
… tests/model-test/cpp_graph_inference.sh