## Type of Change

Stability.ai open sourced some series of StableLM models in 1.6B (base & instructed versions), and one in 3B. This PR adds support for these 3 models.

## Description

- [StableLM2-1.6B](https://huggingface.co/stabilityai/stablelm-2-1_6b)
- [StableLM2-Zephyr-1.6B](https://huggingface.co/stabilityai/stablelm-2-zephyr-1_6b)
- [StableLM-3B](https://huggingface.co/stabilityai/stablelm-3b-4e1t)
- Technical report of 1.6B version: https://arxiv.org/abs/2402.17834

### Features:

- convert_stablelm.py for stablelm2_1.6b/zephyr_1.6b&stablelm3b
- Adds possibility for the user to choose between NE or GGUF format during the convert process
- Model loading for stablelm2_1.6b/zephyr_1.6b&stablelm3b
- Inference function for stablelm2_1.6b/zephyr_1.6b&stablelm3b
- Quant for stablelm2_1.6b/zephyr_1.6b&stablelm3b
- Python bindings for stablelm2_1.6b/zephyr_1.6b&stablelm3b
- Extension test for stablelm2_1.6b
- Update list of supported models


## How has this PR been tested?

- This PR was tested for different settings:

1. **Stablelm-3B** (Conversion NE + Quantization + Inference):

```
(neural-speed) [user1@dnp-test3 neural-speed]$ NEURAL_SPEED_VERBOSE=1 python scripts/run.py stabilityai/stablelm-3b-4e1t --format NE --group_size 128 --compute_dtype int8 --weight_dtype int4 -n 512 -p "Building a website can be done in 10 simple steps:"
Fetching 10 files: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 189787.51it/s]
Convert model ...
cmd: ['python', PosixPath('/home/user1/anaconda3/envs/neural-speed/lib/python3.9/site-packages/neural_speed/convert/convert_stablelm.py'), '--outfile', PosixPath('stablelm_files/ne_stablelm_f32.bin'), '--outtype', 'f32', '--format', 'NE', '/home/user1/.cache/huggingface/hub/models--stabilityai--stablelm-3b-4e1t/snapshots/77fd07d51bbadf7735f52c0d01598ed3e7db1218']
Special tokens have been added in the vocabulary, make sure the associated word embeddings are fine-tuned or trained.
Loading model:  /home/user1/.cache/huggingface/hub/models--stabilityai--stablelm-3b-4e1t/snapshots/77fd07d51bbadf7735f52c0d01598ed3e7db1218
Model loaded:  /home/user1/.cache/huggingface/hub/models--stabilityai--stablelm-3b-4e1t/snapshots/77fd07d51bbadf7735f52c0d01598ed3e7db1218
{'vocab_size': 50304, 'max_position_embeddings': 4096, 'hidden_size': 2560, 'intermediate_size': 6912, 'num_hidden_layers': 32, 'num_attention_heads': 32, 'num_key_value_heads': 32, 'hidden_act': 'silu', 'initializer_range': 0.02, 'layer_norm_eps': 1e-05, 'use_cache': True, 'rope_theta': 10000, 'rope_scaling': None, 'use_qkv_bias': False, 'hidden_dropout': 0.0, 'attention_dropout': 0.0, 'partial_rotary_factor': 0.25, 'return_dict': True, 'output_hidden_states': False, 'output_attentions': False, 'torchscript': False, 'torch_dtype': 'bfloat16', 'use_bfloat16': False, 'tf_legacy_loss': False, 'pruned_heads': {}, 'tie_word_embeddings': False, 'chunk_size_feed_forward': 0, 'is_encoder_decoder': False, 'is_decoder': False, 'cross_attention_hidden_size': None, 'add_cross_attention': False, 'tie_encoder_decoder': False, 'max_length': 20, 'min_length': 0, 'do_sample': False, 'early_stopping': False, 'num_beams': 1, 'num_beam_groups': 1, 'diversity_penalty': 0.0, 'temperature': 1.0, 'top_k': 50, 'top_p': 1.0, 'typical_p': 1.0, 'repetition_penalty': 1.0, 'length_penalty': 1.0, 'no_repeat_ngram_size': 0, 'encoder_no_repeat_ngram_size': 0, 'bad_words_ids': None, 'num_return_sequences': 1, 'output_scores': False, 'return_dict_in_generate': False, 'forced_bos_token_id': None, 'forced_eos_token_id': None, 'remove_invalid_values': False, 'exponential_decay_length_penalty': None, 'suppress_tokens': None, 'begin_suppress_tokens': None, 'architectures': ['StableLmForCausalLM'], 'finetuning_task': None, 'id2label': {0: 'LABEL_0', 1: 'LABEL_1'}, 'label2id': {'LABEL_0': 0, 'LABEL_1': 1}, 'tokenizer_class': None, 'prefix': None, 'bos_token_id': 0, 'pad_token_id': None, 'eos_token_id': 0, 'sep_token_id': None, 'decoder_start_token_id': None, 'task_specific_params': None, 'problem_type': None, '_name_or_path': '/home/user1/.cache/huggingface/hub/models--stabilityai--stablelm-3b-4e1t/snapshots/77fd07d51bbadf7735f52c0d01598ed3e7db1218', 'transformers_version': '4.38.2', 'model_type': 'stablelm'}
model.embed_tokens.weight  ->  model.embed_tokens.weight
model.embed_tokens.weight 2 (50304, 2560)
  Converting to float32 (50304, 2560) [[0.00043487548828125, 0.00092315673828125, 0.000705718994140625], [4.363059997558594e-05, -0.00011396408081054688, 5.289912223815918e-07], [-0.01190185546875, 0.0034027099609375, 0.006927490234375]]
b'model.embed_tokens.weight'
model.layers.0.self_attn.q_proj.weight  ->  model.layers.0.self_attn.q_proj.weight
model.layers.0.self_attn.q_proj.weight 2 (2560, 2560)
  Converting to float32 (2560, 2560) [[-0.00095367431640625, 0.000156402587890625, 0.00061798095703125], [0.0009765625, -8.869171142578125e-05, -0.000553131103515625], [-0.000614166259765625, 0.000152587890625, 0.000263214111328125]]
...
model.layers.1.input_layernorm.weight 1 (2560,)
  Converting to float32 (2560,) [0.146484375, 0.1552734375, 0.185546875]
b'model.layers.1.input_layernorm.weight'
model.layers.1.input_layernorm.bias  ->  model.layers.1.input_layernorm.bias
model.layers.1.input_layernorm.bias 1 (2560,)
...
  Converting to float32 (2560,) [-0.01531982421875, -0.0869140625, 0.130859375]
b'model.norm.bias'
lm_head.weight  ->  lm_head.weight
lm_head.weight 2 (50304, 2560)
  Converting to float32 (50304, 2560) [[0.009765625, -0.01806640625, 0.004180908203125], [0.000545501708984375, 0.00115203857421875, -0.0004558563232421875], [0.01422119140625, 0.0025634765625, 0.00933837890625]]
b'lm_head.weight'
Done. Output file: stablelm_files/ne_stablelm_f32.bin

Quantize model ...
[PosixPath('/home/user1/anaconda3/envs/neural-speed/lib/python3.9/site-packages/neural_speed/quant_stablelm'), '--model_file', PosixPath('stablelm_files/ne_stablelm_f32.bin'), '--out_file', PosixPath('stablelm_files/ne_stablelm_int4.bin'), '--nthread', '1', '--weight_dtype', 'int4', '--alg', 'sym', '--group_size', '128', '--scale_dtype', 'fp32', '--compute_dtype', 'int8']
AVX:1 AVX2:1 AVX512F:1 AVX_VNNI:1 AVX512_VNNI:1 AMX_INT8:1 AMX_BF16:1 AVX512_BF16:1 AVX512_FP16:1
ne_ftype: 10
model.cpp: loading model from stablelm_files/ne_stablelm_f32.bin
Loading the bin file with NE format...
model.cpp: saving model to stablelm_files/ne_stablelm_int4.bin
[   1/ 356]            model.embed_tokens.weight -     2560 x 50304, type =    f32, 0_0_32_0_0,quantizing .. GGML size =   491.25 MB ->    69.08 MB
...
[ 355/ 356]                      model.norm.bias -             2560, type =    f32, 7_0_32_0_0,size =    0.010 MB

[ 356/ 356]                       lm_head.weight -     2560 x 50304, type =    f32, 0_0_128_1_1,quantizing .. JBLAS size =   491.25 MB ->    67.16 MB
...
Welcome to use the stablelm on the ITREX!
main: seed  = 1709743807
AVX:1 AVX2:1 AVX512F:1 AVX_VNNI:1 AVX512_VNNI:1 AMX_INT8:1 AMX_BF16:1 AVX512_BF16:1 AVX512_FP16:1
model.cpp: loading model from stablelm_files/ne_stablelm_int4.bin
Loading the bin file with NE format...
init: n_vocab                  = 50304
init: n_embd                   = 2560
init: n_head                   = 32
init: n_layer                  = 32
init: n_ff                     = 6912
init: n_parts                  = 1
init: n_embd      = 2560
init: max_seq_len      = 4096
load: ne ctx size = 1470.29 MB
load: mem required  = 4544.29 MB (+ memory per state)
.............................................................................................
model_init_from_file: support_bestla_kv = 1
model_init_from_file: kv self size = 1560.00 MB

system_info: n_threads = 56 / 208 | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | F16C = 1 | BLAS = 0 | SSE3 = 1 | VSX = 0 |
sampling: repeat_last_n = 64, repeat_penalty = 1.100000, presence_penalty = 0.000000, frequency_penalty = 0.000000, top_k = 40, tfs_z = 1.000000, top_p = 0.950000, typical_p = 1.000000, temp = 0.800000, mirostat = 0, mirostat_lr = 0.100000, mirostat_ent = 5.000000
generate: n_ctx = 512, n_batch = 512, n_predict = 512, n_keep = 0


Building a website can be done in 10 simple steps:
- Decide what you want to build (web page or blog), and for whom. Juxtapose this against your budget.
- Choose an online web hosting company, preferably located locally or with good Internet bandwidth – such as one like the following:
You need a domain name registered in your name. There are many names available. I have a personal domain called “sjw-dance.com”. The.net is also available and sometimes cheaper for a new website since there’s no competition out there yet. Choose one that suits you best.
- Decide what kind of web site you want to build. You can choose the popular content management system (CMS) as your building tool. These include WordPress, Joomla or Drupal (the latter two are paid). Or you may opt for a more simple approach like Google’s Website Builder which comes with free hosting and is easy to set up.
- Choose the design of the site – if you want one. A professional web designer can help here. If not, use a pre-built template from Themeforest (or similar) which are free or paid. In this latter case, consider your budget. We all like our sites with a custom feel and layout but this may be expensive in the long run if you want it for many years to come.
- Uploading of content – pictures, videos or other media – is best uploaded through an external system that will not disrupt the speed of your site (such as FTP).
- Ensure that all images are named with a relevant keyword and small file size to reduce loading time on a client device. You can do this manually but using online tools is better so you get some automated help instead of having to know every single detail yourself – such as:
- Image Compression Tool for Websites
- Free Image Compressor Online
- Photoshop CC (not essential if you are a pro!)
- Adobe Image Editor
- A text editor. This can be Microsoft Word, Pages/Numbers, LibreOffice Writer or Gedit. Note that some of these editors have an ‘image’ function as part of the package and this can save you time here too. Otherwise there is an excellent web solution called TinyMCE which is a free online image editor at:
- Uploaded content such as images etc will now need to be resized for optimum load times on a client device. Use a photo Editor also available in Microsoft word
Adobe Photoshop
model_print_timings:        load time =    23.81 ms
model_print_timings:      sample time =   320.55 ms /   512 runs   (    0.63 ms per token)
model_print_timings: prompt eval time =   131.94 ms /   268 tokens (    0.49 ms per token)
model_print_timings:        eval time =  8260.31 ms /   510 runs   (   16.20 ms per token)
model_print_timings:       total time =  8873.70 ms
```

2. **Stablelm2-Zephyr-1.6B** (Conversion NE + Quantization + Inference):

```

(neural-speed) [user1@dnp-test3 neural-speed]$ NEURAL_SPEED_VERBOSE=1 python scripts/run.py stabilityai/stablelm-2-zephyr-1_6b --format NE --group_size 128 --compute_dtype int8 --weight_dtype int4 -p "Building a website can be done in 10 simple steps:"
Fetching 20 files: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████| 20/20 [00:00<00:00, 308404.71it/s]
Convert model ...
cmd: ['python', PosixPath('/home/user1/anaconda3/envs/neural-speed/lib/python3.9/site-packages/neural_speed/convert/convert_stablelm.py'), '--outfile', PosixPath('stablelm_files/ne_stablelm_f32.bin'), '--outtype', '
model.embed_tokens.weight  ->  model.embed_tokens.weight
model.embed_tokens.weight 2 (100352, 2048)
  Converting to float32 (100352, 2048) [[0.02734375, 0.0322265625, 0.034912109375], [0.01043701171875, 0.00102996826171875, 0.00061798095703125], [-0.0174560546875, -0.01068115234375, -0.00897216796875]]
b'model.embed_tokens.weight'
model.layers.0.self_attn.q_proj.weight  ->  model.layers.0.self_attn.q_proj.weight
model.layers.0.self_attn.q_proj.weight 2 (2048, 2048)
  Converting to float32 (2048, 2048) [[0.01239013671875, 0.00885009765625, 0.05810546875], [-0.0087890625, -0.00244140625, -0.035888671875], [-0.00823974609375, -0.000331878662109375, -0.0035552978515625]]
b'model.layers.0.self_attn.q_proj.weight'
model.layers.0.self_attn.q_proj.bias  ->  model.layers.0.self_attn.q_proj.bias
model.layers.0.self_attn.q_proj.bias 1 (2048,)
  Converting to float32 (2048,) [0.08154296875, -0.05712890625, -7.375]

...
b'model.layers.23.post_attention_layernorm.weight'
model.layers.23.post_attention_layernorm.bias  ->  model.layers.23.post_attention_layernorm.bias
model.layers.23.post_attention_layernorm.bias 1 (2048,)
  Converting to float32 (2048,) [-0.000255584716796875, -0.0003643035888671875, -0.001220703125]
b'model.layers.23.post_attention_layernorm.bias'
model.norm.weight  ->  model.norm.weight
model.norm.weight 1 (2048,)
  Converting to float32 (2048,) [4.6875, 4.59375, 5.46875]
b'model.norm.weight'
model.norm.bias  ->  model.norm.bias
model.norm.bias 1 (2048,)
  Converting to float32 (2048,) [0.002777099609375, 0.05322265625, -0.09619140625]
b'model.norm.bias'
lm_head.weight  ->  lm_head.weight
lm_head.weight 2 (100352, 2048)
  Converting to float32 (100352, 2048) [[-0.00173187255859375, -0.0029449462890625, 0.00909423828125], [0.006317138671875, -0.00408935546875, 0.01544189453125], [-0.00323486328125, -0.0106201171875, 0.01483154296875]]
b'lm_head.weight'
Done. Output file: stablelm_files/ne_stablelm_f32.bin

Quantize model ...
[PosixPath('/home/user1/anaconda3/envs/neural-speed/lib/python3.9/site-packages/neural_speed/quant_stablelm'), '--model_file', PosixPath('stablelm_files/ne_stablelm_f32.bin'), '--out_file', PosixPath('stablelm_files/ne_stablelm_int4.bin'), '--nthread', '1', '--weight_dtype', 'int4', '--alg', 'sym', '--group_size', '128', '--scale_dtype', 'fp32', '--compute_dtype', 'int8']
AVX:1 AVX2:1 AVX512F:1 AVX_VNNI:1 AVX512_VNNI:1 AMX_INT8:1 AMX_BF16:1 AVX512_BF16:1 AVX512_FP16:1
ne_ftype: 10
model.cpp: loading model from stablelm_files/ne_stablelm_f32.bin
Loading the bin file with NE format...
model.cpp: saving model to stablelm_files/ne_stablelm_int4.bin
[   1/ 340]            model.embed_tokens.weight -    2048 x 100352, type =    f32, 0_0_32_0_0,quantizing .. GGML size =   784.00 MB ->   110.25 MB

[   2/ 340] model.layers.0.self_attn.q_proj.weight -     2048 x  2048, type =    f32, 0_0_128_1_1,quantizing .. JBLAS size =    16.00 MB ->     2.20 MB

...
[ 339/ 340]                      model.norm.bias -             2048, type =    f32, 7_0_32_0_0,size =    0.008 MB

[ 340/ 340]                       lm_head.weight -    2048 x 100352, type =    f32, 0_0_128_1_1,quantizing .. JBLAS size =   784.00 MB ->   107.20 MB

model_quantize_internal: model size  =  6273.33 MB
model_quantize_internal: quant size  =   866.37 MB

main: quantize time = 24944.99 ms
main:    total time = 24944.99 ms
Inference model ...
Namespace(model_name='stablelm', model=PosixPath('stablelm_files/ne_stablelm_int4.bin'), build_dir=PosixPath('/home/user1/aahouzi/neural-speed/scripts/../build'), prompt='Building a website can be done in 10 simple steps:', tokenizer='THUDM/chatglm-6b', n_predict=-1, threads=56, batch_size_truncate=512, ctx_size=512, seed=-1, repeat_penalty=1.1, color=False, keep=0, shift_roped_k=False, memory_f32=False, memory_f16=False, memory_auto=False, one_click_run='True')
cmd: [PosixPath('/home/user1/anaconda3/envs/neural-speed/lib/python3.9/site-packages/neural_speed/run_stablelm'), '--model', PosixPath('stablelm_files/ne_stablelm_int4.bin'), '--prompt', 'Building a website can be done in 10 simple steps:', '--n-predict', '-1', '--threads', '56', '--batch-size-truncate', '512', '--ctx-size', '512', '--seed', '-1', '--repeat-penalty', '1.1', '--keep', '0']
Welcome to use the stablelm on the ITREX!
main: seed  = 1709751627
AVX:1 AVX2:1 AVX512F:1 AVX_VNNI:1 AVX512_VNNI:1 AMX_INT8:1 AMX_BF16:1 AVX512_BF16:1 AVX512_FP16:1
model.cpp: loading model from stablelm_files/ne_stablelm_int4.bin
Loading the bin file with NE format...
init: n_vocab                  = 100352
init: n_embd                   = 2048
init: n_head                   = 32
init: n_layer                  = 24
init: n_ff                     = 5632
init: n_parts                  = 1
init: n_embd      = 2048
init: max_seq_len      = 4096
load: ne ctx size =  866.47 MB
load: mem required  = 2916.47 MB (+ memory per state)
.............................................................................
model_init_from_file: support_bestla_kv = 1
model_init_from_file: kv self size =  972.00 MB

system_info: n_threads = 56 / 208 | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | F16C = 1 | BLAS = 0 | SSE3 = 1 | VSX = 0 |
sampling: repeat_last_n = 64, repeat_penalty = 1.100000, presence_penalty = 0.000000, frequency_penalty = 0.000000, top_k = 40, tfs_z = 1.000000, top_p = 0.950000, typical_p = 1.000000, temp = 0.800000, mirostat = 0, mirostat_lr = 0.100000, mirostat_ent = 5.000000
generate: n_ctx = 512, n_batch = 512, n_predict = -1, n_keep = 0


Building a website can be done in 10 simple steps: First, choose a domain name. Second, select the platform for your website (such as WordPress or Wix). Third, create your content and design with an intuitive layout. Fourth, install any necessary plugins or add-ons to optimize your site's performance. Finally, promote your website through social media, email marketing campaigns, or paid advertising.

Here are some more specific steps:

1. Choose a domain name: This is the first step in building your website. You'll need something memorable, easy to spell, and available for registration.

2. Select your platform: Depending on your preferences and budget, you can choose from various platforms like WordPress, Wix, or Squarespace.

3. Plan your content: Think about what kind of information, images, and videos will be featured on your website. A clear outline should guide your content creation process.

4. Design your website with an intuitive layout: Make sure to use a clean and clear design that reflects your brand identity while maintaining easy navigation for visitors.

5. Install any necessary plugins or add-ons: To optimize your site's performance, install essential plugins like Google Analytics, security features, or social media integrations.

6. Promote your website through marketing channels: Once you've launched your website, continue to promote it through email campaigns, social media marketing, paid advertising, and other promotional techniques.

Remember that building a successful website requires ongoing maintenance and continuous optimization efforts. By following these steps, you'll be well on your way to developing a professional-looking website with high-quality content and an intuitive layout.<|endoftext|> [end of text]

model_print_timings:        load time =    16.76 ms
model_print_timings:      sample time =   389.78 ms /   314 runs   (    1.24 ms per token)
model_print_timings: prompt eval time =    16.74 ms /    12 tokens (    1.40 ms per token)
model_print_timings:        eval time =  2868.66 ms /   313 runs   (    9.17 ms per token)
model_print_timings:       total time =  3465.19 ms
```

3. **Stablelm2-1.6B** (Conversion NE + Quantization + Inference):

```
(neural-speed) [user1@dnp-test3 neural-speed]$ NEURAL_SPEED_VERBOSE=1 python scripts/run.py stabilityai/stablelm-2-1_6b --format NE --group_size 128 --compute_dtype int8 --weight_dtype int4 -n 512 -p "Building a website can be done in 10 simple steps:"
Fetching 13 files: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████| 13/13 [00:00<00:00, 207323.01it/s]
…