@a32543254 2024-03-14
@intellinjun  could you help add the extension test for this new models