@intellinjun 2024-03-14
https://inteltf-jenk.sh.intel.com/job/neural_speed_extension/90/
Please wait for the full performance test results