@Zhenzhong1 2024-03-14
> > @intellinjun all checks look good, should be ready for review. On seperate note, any leads on GGUF conversion issue that I mentioned ?
> > @Zhenzhong1 Please help reply.

no worries. we have the issue to track this. https://github.com/intel/neural-speed/issues/94