@aahouzi 2024-03-14
> https://inteltf-jenk.sh.intel.com/job/neural_speed_extension/90/ Please wait for the full performance test results

@intellinjun I see in the link that this is for 1.6B only, can we do the same for 3B too ?

and is it SPR dual or one socket ?